+++ include/ksnp/server.h
@@ -333,7 +333,8 @@ bool ksnp_server_want_read(struct ksnp_server const *server) NOEXCEPT;
  * @param len [in,out] Pointer to the length value, which must match the length
  * of the input buffer @p data. If this function returns successfully, the
  * number of bytes actually read are written to the length value, which may be
- * 0.
+ * less than the input if the buffer of the associated message context does not
+ * have sufficient capacity.
  * @return @ref KSNP_E_NO_ERROR on success.
  * @return Any of the values from the @ref ksnp_error enum on failure.
  */